crud OK! ajout des interface repository

+++ src/main/java/ma/pfa/gestion_evenement/Controller/CompteController.java
@@ -26,9 +26,8 @@ public class CompteController{
     }
 
     @DeleteMapping("/deleteAccount/{id}")
-    public void delete(@PathVariable("id")  Integer id){
+    public void delete(@PathVariable("id") Long id){
         compteRepository.deleteById(id);
-
     }
 
     @PutMapping("/update")
