+++ src/main/java/ma/pfa/gestion_evenement/Repository/AdministrateurRepository.java
@@ -3,6 +3,6 @@ package ma.pfa.gestion_evenement.Repository;
 import ma.pfa.gestion_evenement.entity.Administrateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdministrateurRepository extends JpaRepository <Administrateur,Integer> {
+public interface AdministrateurRepository extends JpaRepository <Administrateur,Long> {
 
 }
