+++ src/main/java/ma/pfa/gestion_evenement/Repository/CompteRepository.java
@@ -3,6 +3,6 @@ package ma.pfa.gestion_evenement.Repository;
 import ma.pfa.gestion_evenement.entity.Compte;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompteRepository extends JpaRepository<Compte,Integer> {
+public interface CompteRepository extends JpaRepository<Compte,Long> {
 
 }
