+++ src/main/java/ma/pfa/gestion_evenement/Repository/DirecteurRepository.java
@@ -3,5 +3,5 @@ package ma.pfa.gestion_evenement.Repository;
 import ma.pfa.gestion_evenement.entity.Directeur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DirecteurRepository extends JpaRepository<Directeur,Integer> {
+public interface DirecteurRepository extends JpaRepository<Directeur,Long> {
 }
