+++ src/main/java/ma/pfa/gestion_evenement/Repository/EventRepository.java
@@ -3,5 +3,5 @@ package ma.pfa.gestion_evenement.Repository;
 import ma.pfa.gestion_evenement.entity.Compte;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventRepository extends JpaRepository<Compte,Integer> {
+public interface EventRepository extends JpaRepository<Compte,Long> {
 }
