+++ src/main/java/ma/pfa/gestion_evenement/Repository/OrganisateurRepository.java
@@ -3,5 +3,5 @@ package ma.pfa.gestion_evenement.Repository;
 import ma.pfa.gestion_evenement.entity.Organisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganisateurRepository extends JpaRepository<Organisateur,Integer> {
+public interface OrganisateurRepository extends JpaRepository<Organisateur,Long> {
 }
