+++ src/main/java/ma/pfa/gestion_evenement/Repository/RespoLogistiqueRepository.java
@@ -3,5 +3,5 @@ package ma.pfa.gestion_evenement.Repository;
 import ma.pfa.gestion_evenement.entity.RespoLogistique;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RespoLogistiqueRepository extends JpaRepository<RespoLogistique,Integer> {
+public interface RespoLogistiqueRepository extends JpaRepository<RespoLogistique,Long> {
 }
