+++ src/main/java/ma/pfa/gestion_evenement/Repository/SalleRepository.java
@@ -3,7 +3,7 @@ package ma.pfa.gestion_evenement.Repository;
 import ma.pfa.gestion_evenement.entity.Salle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SalleRepository extends JpaRepository<Salle,Integer> {
+public interface SalleRepository extends JpaRepository<Salle,Long> {
 
 
 }
